refactor!: move data into json
and deprecate imgToMap
