+++ README.md
@@ -12,9 +12,9 @@
 Если вы хотите добавить картинки, сначала поместите их в директорию `public` и закоммитьте это. Сообщение коммита должно начинаться с `data: `.
 
 Каждая карта имеет фоновое изображение и может иметь точки (`entries`).
-Каждая точка имеет заголовок, описание и может содержать одну или несколько карт (размещенных внутри `children`). Если вы хотите поместить туда изображение вместо карты, оберните url функцией `imgToMap`, вот пример:
+Каждая точка имеет заголовок, описание и может содержать одну или несколько карт (размещенных внутри `children`). Если вы хотите использовать изображение вместо карты, поместите его URL в двойные кавычки, к примеру:
 ```ts
-imgToMap("https://upload.wikimedia.org/wikipedia/commons/4/44/Europe_orthographic_Caucasus_Urals_boundary_%28with_borders%29.svg")
+"https://upload.wikimedia.org/wikipedia/commons/4/44/Europe_orthographic_Caucasus_Urals_boundary_%28with_borders%29.svg"
 ```
 
 Всё, кроме `children` (вложенных карт или изображений), может быть сгенерировано с помощью [графического редактора](https://alexanderthesensei.github.io/map/edit).
